Breaking changes! Changed the way DynaDoc is required and clients are created. Singleton factory is now being used to create clients for models. Added a way for the user to pull the Joi object out of DynaDoc to create the schema instead of requiring them to use Joi. This will require a rewrite of the API once as well.
