+++ lib/dymodel/dymodel.js
@@ -262,6 +262,39 @@ DyModel.prototype.init = function init() {
     this.createTablePayload.TableName = this.modelName;
 }
 
+/**
+The Joi Validate method against this DyModel object.
+@param item (Object): The item to validate against this model.
+@param options (object): Joi Options for validate.
+@param callback (function): Function to callback from the Joi validation.
+@return JoiResult (Object): Joi object that has both a error and value
+  property. If error is defined, then there was an error. Value is the
+  valid object.
+**/
+DyModel.prototype.validate = function validate(item, options, callback) {
+    return Joi.validate(item, this.model, options, callback);
+}
+
+/**
+The Joi assert method. Throws validation error if validation fails.
+Returns nothing.
+@param item (Object): The item to validate against this model.
+@param message (String):  Optional message to display if validation fails.
+**/
+DyModel.prototype.assert = function assert(item, message) {
+    Joi.assert(item, this.model, message);
+}
+
+/**
+The Joi attempt method. Throws validation error if validation fails.
+Also returns the valid object.
+@param item (Object): The item to validate against this model.
+@param message (String):  Optional message to display if validation fails.
+@return item (object): The valid object that passed validation.
+**/
+DyModel.prototype.attempt = function attempt(item, message) {
+    return Joi.attempt(item, this.model, message);
+}
 
 
 module.exports = DyModel;
