+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "dynadoc",
-    "version": "0.3.1",
+    "version": "0.4.0",
     "description": "An easier and intelligent way to use DynamoDB.",
     "main": "dynadoc.js",
     "scripts": {
